Add controllers' tests (#18)
* Add extra field to Book entity class

* BookController - if list is empty return 200 OK

* Return 406 NOT_ACCEPTABLE if book is not available

* Fix PUT replaceBook() method in BookService

* Add test properties

* Add dependency to execute test PATCH endpoints

* Add Book json tests

* Add checkouts list and reservations list to user

* Change relation @ManyToOne to @OneToOne in Reservation
For now assumed that a book can be reserved by only
one user at time. There's no waiting list.

* Update GET /users endpoint
If a returned list of users is empty, a response is
200 OK (body is empty).

* Update GET /users/{id} endpoint

* Update DELETE /users/{id} endpoint

* Update PATCH /users/{id} endpoint

* Update SecurityFilterChain

* Update GET /checkouts endpoint

* Update GET /checkouts/{id} endpoint

* Update GET /reservations endpoint

* Update POST /reservations endpoint

* Update POST /reservations endpoint

* Update DELETE /reservations{id} endpoint

* Add book controller tests

* Add user controller tests

* Add checkout controller tests

* Add reservation controller tests

* Add auth controller tests

* Add register controller tests

* Update README.md

+++ src/main/java/com/example/libraryapp/web/UserController.java
@@ -1,12 +1,16 @@
 package com.example.libraryapp.web;
 
+import com.example.libraryapp.domain.config.CustomSecurityConfig;
 import com.example.libraryapp.domain.config.assembler.UserModelAssembler;
+import com.example.libraryapp.domain.exception.CannotUpdateUserDataException;
+import com.example.libraryapp.domain.exception.UserHasNotReturnedBooksException;
 import com.example.libraryapp.domain.exception.UserNotFoundException;
 import com.example.libraryapp.domain.user.UserService;
 import com.example.libraryapp.domain.user.dto.UserDto;
 import com.example.libraryapp.domain.user.dto.UserUpdateDto;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -27,39 +31,63 @@ public class UserController {
     @GetMapping("/users")
     public ResponseEntity<CollectionModel<EntityModel<UserDto>>> getAllUsers() {
         List<UserDto> allUsers = userService.findAllUsers();
-        if (allUsers.isEmpty()) {
-            return ResponseEntity.noContent().build();
-        }
         CollectionModel<EntityModel<UserDto>> collectionModel = userModelAssembler.toCollectionModel(allUsers);
         return ResponseEntity.ok(collectionModel);
     }
 
     @GetMapping("/users/{id}")
     public ResponseEntity<EntityModel<UserDto>> getUserById(@PathVariable Long id) {
-        return userService.findUserById(id)
-                .map(userModelAssembler::toModel)
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+        try {
+            boolean requestFromOwner = userService.getCurrentLoggedInUserId() == id;
+            boolean requestFromAdmin = userService.getCurrentLoggedInUserRole().equals(CustomSecurityConfig.ADMIN_ROLE);
+
+            if (requestFromOwner || requestFromAdmin) {
+                return userService.findUserById(id)
+                        .map(userModelAssembler::toModel)
+                        .map(ResponseEntity::ok)
+                        .orElseGet(() -> ResponseEntity.notFound().build());
+            } else throw new UserNotFoundException();
+        } catch (UserNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/users/{id}")
     public ResponseEntity<?> deleteUserById(@PathVariable Long id) {
         try {
-            userService.deleteUserById(id);
-            return ResponseEntity.noContent().build();
+            boolean requestFromOwner = userService.getCurrentLoggedInUserId() == id;
+            boolean requestFromAdmin = userService.getCurrentLoggedInUserRole().equals(CustomSecurityConfig.ADMIN_ROLE);
+
+            if (requestFromOwner || requestFromAdmin) {
+                userService.deleteUserById(id);
+                return ResponseEntity.noContent().build();
+            } else throw new CannotUpdateUserDataException();
         } catch (UserNotFoundException e) {
             return ResponseEntity.notFound().build();
+        } catch (CannotUpdateUserDataException e) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        } catch (UserHasNotReturnedBooksException e) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE)
+                    .header("Reason", "User's books are not returned.")
+                    .build();
         }
     }
 
     @PatchMapping("/users/{id}")
     public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody UserUpdateDto user) {
         try {
-            UserDto updatedUser = userService.updateUser(id, user);
-            EntityModel<UserDto> entityModel = userModelAssembler.toModel(updatedUser);
-            return ResponseEntity.ok(entityModel);
+            boolean isDataOwner = userService.getCurrentLoggedInUserId() == id;
+            boolean isAdmin = userService.getCurrentLoggedInUserRole().equals(CustomSecurityConfig.ADMIN_ROLE);
+
+            if (isDataOwner || isAdmin) {
+                UserDto updatedUser = userService.updateUser(id, user);
+                EntityModel<UserDto> entityModel = userModelAssembler.toModel(updatedUser);
+                return ResponseEntity.ok(entityModel);
+            } else throw new CannotUpdateUserDataException();
         } catch (UserNotFoundException | NullPointerException e) {
             return ResponseEntity.notFound().build();
+        } catch (CannotUpdateUserDataException e) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
     }
 }
